requirements.txt
fastapi
uvicorn
sqlalchemy
streamlit
pandas
transformers
torch
accelerate
bitsandbytes
python-dotenv
plotly
pydantic
aiosqlite
sqlparse
numpy
ctransformers[cuda]
tqdm
requests
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.13